+++ src/Display.hpp
@@ -25,7 +25,7 @@ class Display
 		void print(void);
 
 	private:
-		std::vector<int> samplesToPixels(const std::vector<double>& samples, double offset, double verticalScale, int pixelsPerDivision);
+		std::vector<int> samplesToPixels(const std::vector<float>& samples, float offset, float verticalScale, int pixelsPerDivision);
 		void clearScreen(void);
 		void updateScreen(void);
 		void printGrid(const std::vector<int>& coordenates);
@@ -38,7 +38,7 @@ class Display
 		void printButton(std::string title, std::string text, bool isActive, const std::vector<int>& coordenates);
 		void printMenu(const std::vector<std::string>& options, const std::vector<int>& coordenates);
 		void printAlert(std::string text, const std::vector<int>& coordenates);
-		void printSamples(const std::vector<int>& samples, int delay, int memoryDepth, double step, int color, const std::vector<int>& coordenates);
+		void printSamples(const std::vector<int>& samples, int delay, int memoryDepth, float step, int color, const std::vector<int>& coordenates);
 		void printCursor(const std::vector<int>& coordenates);
 		void printMouse(int x, int y);
 };
